+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.0] - 2026-07-14
+
 ### Added
 
 - Opt-in Project Link roles, stack tags, and direct peer relationships, with